Moved the Request Validation logic into a separate Swift file.
